requirements.txt
dnspython==1.16.0
eventlet==0.30.2
Flask==2.0.1
Flask-Login==0.5.0
Flask-SocketIO==4.3.1
gunicorn==20.1.0
igraph==0.9.8
Jinja2==3.0.1
Pillow==8.4.0
psycopg2-binary==2.9.2
pycairo==1.20.1
pymongo==3.12.0
python-dateutil==2.8.2
python-engineio==3.13.2
python-socketio==4.6.0
pytz==2021.1
virtualenv==20.8.1
virtualenvwrapper-win==1.2.6
Werkzeug==2.0.1
